Implement tag balancing for the HTML postprocessor
All open tags should be closed, in the right order (if no closing tag,
add it). A closing tag must close an open tag (if not, remove it).

Of course, the tags in "<br> and </br>" are not open/close tags, so
they will be automagically converted to: "<br /> and <br />".